Resolve failing tests
While fixing the Keyboard `refs` critique, I discovered that not all
languages and layouts were likely being treaded correctly (eg. the
regions that are now identified as 'Standard' formerly would have been
overwritten, or the reverse, by regionally named variants (eg. `Arabic`
vs. `Arabic (Iraq)`) Since I was assuming one without a named region
didn't have any regional variants.

+++ tests/run_tests.sh
@@ -1,8 +1,10 @@
 #!/usr/bin/env bash
 
-apt-get update
-apt install --assume-yes --no-install-recommends --no-install-suggests \
-  libperl-critic-perl \
-  libtest2-suite-perl
+if [ "$(dpkg -l | grep -P 'lib(perl-critic|test2-suite)-perl' | wc -l)" -lt 2 ]; then
+  apt-get update
+  apt install --assume-yes --no-install-recommends --no-install-suggests \
+    libperl-critic-perl \
+    libtest2-suite-perl
+fi
 
 prove /usr/spamtagger/tests/
